MAINT: more rearranging
