+++ env.yaml
@@ -13,6 +13,7 @@ dependencies:
   - scipy
   - pytest
   - python-snappy
+  - sphinx
   - pip:
     - fastapi[all]
     - uvicorn
@@ -27,5 +28,7 @@ dependencies:
     - joblib
     - httpx
     - starlette-prometheus
+    - numpydoc
+    - sphinx_rtd_theme
 prefix: /opt/conda
