+++ salmon/triplets/algs/__init__.py
@@ -1,3 +1,2 @@
 from ._random_sampling import RandomSampling
 from ._round_robin import RoundRobin
-from .utils import Runner
